Decrease module load time
When dot-sourcing script files, PowerShell attempts to validate each
file as you read it in. The alternative is to dot-source a script block
that contains the contents of the file(s).

In my testing of this module I was able to save between 7 and 9 seconds.

Get-Module | Remove-Module -Force
Measure-Command{Import-Module Rapid7Nexpose}